Simplify Importing
I've moved `bootstrap.py` to `__init__.py`; this means that rather than
requiring `from bayesian_bootstrap.bootstrap import X` for imports we
only need to say `from bayesian_bootstrap import X`, simplifying things
a little. I've updated the README, demos, and tests to reflect this
change. I also tweaked the delta value for one test, as it failed
intermittently with randomized data.
